Modified and extended ROpenCL for the book
enqueueNDRangeKernel() now takes localWorkSize as an optional parameter
so you don’t _have_ to specify it, in which case the OpenCL driver will
choose an appropriate number. With OpenCL as it currently stands <2.0
you still have to choose globalWorkSize as a multiple of localWorkSize,
so stick to multiples of 1024 and you should be safe. Also it can now
take global and localWorkSize as vectors (up to 2D right now) to make
support for 2D matrix algs more natural for R programmers rather than
forcing them always into a 1D world. [TODO complete by adding in
support for 3D.]

enqueueReadBuffer() can now take an optional offset into the R vector
to start copying into. This enables a single R result vector to be
built up through multiple kernel invocations and saves some memory copy
overhead as well as reducing overall memory requirements.

assignKernelArg() added to enable the parameter to a compiled kernel to
be changed between invocations. Called it assign to differentiate from
OpenCL setKernelArg which is already exposed and called internally.
assignKernelArg figures out appropriate setKernelArg to call based on
the R parameter value type.

Added deviceSupports[Single|Double|Half]Precision() convenience
functions to make it easier for R programmers to switch their code
based on the precision capability of the underlying device (returns
TRUE|FALSE). If you pass in a second arg as an empty list(), then this
will be filled out with the details of exactly what floating point
capabilities the device has for this precision (rounding etc).